Watch: makeID mit Int64 statt Int (kein arm64_32-Crash).
Epoch-Millis (~1.7e12) sprengen Int32.max. Auf der Watch (32-bit Int)
crasht jedes Anlegen in Item.makeID — inkl. consumeSiriPendingAdds.
SavedList und StoreCatalog nutzen denselben Int64-Pfad.

+++ Sources/Shared/StoreLayout.swift
@@ -73,7 +73,7 @@ enum StoreCatalog {
     static let customDefaultLayout = ["vor", "sonstiges", "nach"]
 
     static func makeID() -> String {
-        let t = String(Int(Date().timeIntervalSince1970 * 1000), radix: 36)
+        let t = String(Int64(Date().timeIntervalSince1970 * 1000), radix: 36)
         let r = String(UInt64.random(in: 0..<0xFFFFFF), radix: 36)
         return "s\(t)\(r)"
     }
